Infrastructure for context sharing. - Regal shared_ptr, shared_map and shared_list added. - Initial sharing support for RegalObj and RegalVao, not used yet. - RegalContext has a shared_list of other contexts for sharing, not used yet.

+++ src/regal/RegalIff.cpp
@@ -1824,7 +1824,7 @@ void RFF::State::Process( RegalIff * ffn ) {
       }
     }
   }
-  p.hash = compute_hash(p);
+  p.hash = (GLuint) compute_hash(p);
 }
 
 void RFF::UpdateUniforms( RegalContext * ctx ) {
